working on minilog with auth; debug needed

+++ minilog-jpa-with-auth/src/main/java/com/asdf/minilog/service/ArticleService.java
@@ -4,6 +4,7 @@ import com.asdf.minilog.dto.ArticleResponseDto;
 import com.asdf.minilog.entity.Article;
 import com.asdf.minilog.entity.User;
 import com.asdf.minilog.exception.ArticleNotFoundException;
+import com.asdf.minilog.exception.NotAuthorizedException;
 import com.asdf.minilog.exception.UserNotFoundException;
 import com.asdf.minilog.repository.ArticleRepository;
 import com.asdf.minilog.repository.UserRepository;
@@ -14,6 +15,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Transactional;
 
+/** 변경점 JWT 토큰을 이용한 Authorization 구현 deleteArticle, updateArticle */
 @Service
 @Transactional(isolation = Isolation.REPEATABLE_READ)
 public class ArticleService {
@@ -38,12 +40,12 @@ public class ArticleService {
                                                         "해당 아이디(%d)를 가진 사용자를 찾을 수 없습니다.", userId)));
 
         Article article = Article.builder().content(content).author(user).build();
-
         Article savedArticle = articleRepository.save(article);
+
         return EntityDtoMapper.toDto(savedArticle);
     }
 
-    public void deleteArticle(Long articleId) {
+    public void deleteArticle(Long authorId, Long articleId) {
         Article article =
                 articleRepository
                         .findById(articleId)
@@ -53,11 +55,15 @@ public class ArticleService {
                                                 String.format(
                                                         "해당 아이디(%d)를 가진 게시글을 찾을 수 없습니다.",
                                                         articleId)));
+
+        if (!article.getAuthor().getId().equals(authorId)) {
+            throw new NotAuthorizedException("게시글 작성자만 삭제할 수 있습니다.");
+        }
 
         articleRepository.deleteById(articleId);
     }
 
-    public ArticleResponseDto updateArticle(Long articleId, String content) {
+    public ArticleResponseDto updateArticle(Long authorId, Long articleId, String content) {
         Article article =
                 articleRepository
                         .findById(articleId)
@@ -67,6 +73,10 @@ public class ArticleService {
                                                 String.format(
                                                         "해당 아이디(%d)를 가진 게시글을 찾을 수 없습니다.",
                                                         articleId)));
+
+        if (!article.getAuthor().getId().equals(authorId)) {
+            throw new NotAuthorizedException("게시글 작성자만 수정할 수 있습니다.");
+        }
 
         article.setContent(content);
         Article updatedArticle = articleRepository.save(article);
